refactor(theme): complete theme context & hooks
