+++ packages/dumi/theme/index.d.ts
@@ -1,6 +1,1 @@
-export { default as context } from '@umijs/preset-dumi/src/theme/context';
-export { default as SlugLink } from '@umijs/preset-dumi/src/theme/components/SlugLink';
-export { default as NavbarLink } from '@umijs/preset-dumi/src/theme/components/NavbarLink';
-export { default as useCopy } from '@umijs/preset-dumi/src/theme/hooks/useCopy';
-export { default as useSearch } from '@umijs/preset-dumi/src/theme/hooks/useSearch';
-export { default as useCodeSandbox } from '@umijs/preset-dumi/src/theme/hooks/useCodeSandbox';
+export * from '@umijs/preset-dumi/src/theme';
