+++ packages/preset-dumi/src/theme/components/Link.tsx
@@ -1,0 +1,1 @@
+export { Link as default } from '@umijs/runtime';
