+++ packages/preset-dumi/src/theme/index.ts
@@ -1,6 +1,39 @@
+import { IDepAnalyzeResult } from '../transformer/demo/dependencies';
+
 export { default as context } from './context';
-export { default as NavbarLink } from './components/NavbarLink';
+export { default as Link } from './components/Link';
+export { default as NavLink } from './components/NavLink';
 export { default as SlugLink } from './components/SlugLink';
 export { default as useSearch } from './hooks/useSearch';
 export { default as useCopy } from './hooks/useCopy';
 export { default as useCodeSandbox } from './hooks/useCodeSandbox';
+
+export interface IPreviewerComponentProps {
+  title?: string;
+  description?: string;
+  sources: {
+    /**
+     * self source code for demo
+     * @note  jsx exsits definitely, tsx exists when the source code language is tsx
+     */
+    _: { jsx: string; tsx?: string };
+  } & {
+    /**
+     * other file source code which imported in demo
+     */
+    [key: string]: { import: string } & IPreviewerComponentProps['sources']['_'];
+  };
+  /**
+   * third-party dependencies of demo
+   */
+  dependencies: IDepAnalyzeResult['dependencies'];
+  /**
+   * global identifier for demo
+   */
+  identifier: string;
+  /**
+   * the component which demo belongs to
+   */
+  componentName?: string;
+  [key: string]: any;
+}
